Replace old GameState with new one
